hgext: move "jenkins" extension out of jpl module
jpl extension disables Mercurial's demandimport (dunno why) hence making "hg"
command very slow to because it tries to import requests and friends. In fact,
"jenkins" extension does not really need jpl apart from when one wants to rely
on a CubicWeb forge to retrieve the repository URL when configuration is
missing. So move the "jenkins" in top-level "hgext" package to restore normal
startup time for users of the "jenkins" extension.