hgext: add an invalidation mechanism for "jenkins" store
We rely on repository's tip revision to determine if "jenkins" store is valid.
Otherwise, file's data is not used and thus the cache is rebuilt (and saved
afterwards).

This makes it possible to rely on the local state of the repository (that
changes after a pull or a commit) to either use or invalidate the "jenkins"
cache.